+++ server/src/main/java/lsfusion/server/remote/UserActivity.java
@@ -1,10 +1,10 @@
 package lsfusion.server.remote;
 
 public class UserActivity {
-    public Long computer;
+    public Integer computer;
     public Long time;
 
-    public UserActivity(Long computer, Long time) {
+    public UserActivity(Integer computer, Long time) {
         this.computer = computer;
         this.time = time;
     }
